Remove `from_master` inferrers and placeholders.
Add inference for type constructors `object` and None type.

Change test_math_log to test_module_function_call to keep testing inference for a module-imported function.
